IGNITE-13734: Register service return type on method invocation.
