+++ modules/platforms/dotnet/Apache.Ignite.Core.Tests/Services/JavaServiceDynamicProxy.cs
@@ -21,7 +21,7 @@ namespace Apache.Ignite.Core.Tests.Services
     using System.Collections;
     using System.Collections.Generic;
     using Apache.Ignite.Core.Binary;
-    using org.apache.ignite.platform;
+    using org.apache.ignite.platform.model;
 
     /// <summary>
     /// Explicit service proxy over dynamic variable.
@@ -322,6 +322,11 @@ namespace Apache.Ignite.Core.Tests.Services
         public Account[] testAccounts()
         {
             return _svc.testAccounts();
+        }
+
+        public User[] testUsers()
+        {
+            return _svc.testUsers();
         }
 
         /** <inheritDoc /> */
